Split proper-list-p into two implementations
For 26.1 onwards, length throws an error when the list contains a
cycle.  This can be caught, and told to return a nil value.

For older versions, a variation of Richard P. Brent's cycle detection
algorithm is used, before length is called if we make sure that the
list contains no cycles.

Once more, thanks to Mattias Engdegård for suggesting these
improvements.