Compress and merge the compiled result of compat.elc
The previous `compat-generate-common' (now `compat--generate-verbose')
generated a lot of code that was not relevant for 99% of most users,
and can be stripped away, reducing the size of the resulting program
and increasing the speed at which compat is loaded.  These factors are
worth optimising and investing complexity into, as compat autoloads
itself and therefore shouldn't burden the user unnecessarily.