Use export_file_name Org export keyword
Previously the "texinfo_filename" export keyword that was used
instead.  But that only changes the value of the "@setfilename"
command in the texi file but not the actual names of the produced
texi and info files, making it necessary to explicitly rename the
texi file in the Makefile and making it impossible to produce the
info file using "org-export-dispatch".

"export_file_name" doesn't have these problems.