Use Texinfo definition commands and add indices
Use the appropriate Texinfo definition commands for functions, macros
and variables.  This improves how these definitions are displayed in
the Info manual.

To accomplish this, make use of a new (and yet to be released) feature
in Org's main branch, described in (info "(org)Plain lists in Texinfo
export").  Doing that takes care of adding entries to the appropriate
indices.

Also add two new sections to actually display the indices.

Because the required Org version isn't releases yet and thus not
installed yet on the server used to build packages for [Non]GNU Elpa,
we have to check the generated Texinfo file into git.