Do not omit the prefix when documenting a prefixed symbol
Implemented functions and variables appear in the respective indices
exactly as written.  If we do not prefix these symbols then the risk
is very high that someone will look at the index and see that some
function is defined in this package and start to use it, when in
actuality it is only the prefixed symbol that is defined here.