Assume compat-XY defines definitions for Emacs XY.1
This avoids confusion when someone is tracking an older development
release, so that compat does not assume all the functionality that was
part of the final release was already available.  Instead on a XY.0.Z
release, every definition is checked for availability, just like with
any previous release.

Thanks to Jacek Złydach for bringing the issue to my attention:
https://lists.sr.ht/~pkal/compat-devel/%3Cf8635d7d-e233-448f-b325-9e850363241c%40www.fastmail.com%3E