Fix using compat libraries uncompiled
`compat-entwine' is called in two and only three distinct situations.

1. When "compat.el" is being loaded without having been compiled
   before.  This can happen while a third-party library is being
   compiled, which requires `compat'.

2. When "compat.el" is being evaluated, such as when using
   `eval-buffer'.

3. When "compat.el" is being compiled.

In all cases we need to detect where "compat.el" is located, because
that also tells us where the "compat-XY.el" libraries are located and
`compat-entwine' needs to read the contents of those files.  To detect
the location of "compat.el" we consult various variables, which we
expect to either be nil or "/path/to/compat.el".  We didn't use the
correct variables.

Some time ago Stefan added a call to `macro-file-name' to the front
of that search, which fixes the search for Emacs 28.1.  Older Emacs
releases lack that function, so for those we still ended up consulting
the wrong variables.  To fix that use `current-load-list' directly,
treating it the same way as `macro-file-name' does.

If that gives us a non-nil value, then we are dealing with one of the
first two cases and we are done.  Otherwise we are dealing with the
third case and `byte-compile-current-file' gives us the correct value.

One of these sources is always non-nil, so drop the variables we used
to fall back to.

Then we pass the version from `compat-entwine' to the
`compat--generate-function' function by binding the new variable
`compat--entwine-version'.  Previously this was done by binding
`byte-compile-current-file`.  The caller then had to extract the
version from the name of that file.

Binding `byte-compile-current-file` was also part of the bug that
prevented compilation of a third-party library, which depends on
"compat-XY.el", when compat's libraries have not been compiled yet.

To fix that we have to change the order of the possible sources of the
version string in `compat--generate-function' functions.  As before we
use the version explicitly specified for the function being processed,
if any.  Then we use value of the new `compat--entwine-version'
variable, if that is bound and non-nil, i.e., if the binding from
`compat-entwine' is in effect.

Previously we would have tried `byte-compile-current-file' at this
point instead, but while that might be non-nil because we used to bind
it in `compat--generate-function', it could also be non-nil because an
arbitrary third-party library is being compiled.

That is a problem if that third-party library depends on
"compat-XY.el" and the latter hasn't been compiled before: The
`compat--generate-function' function is called directly in that case,
without the binding from `compat-entwine' in effect.  We need
`byte-compile-current-file' to be some "compat-XY.el" library; not
some arbitrary third-party library.

Then we try `current-load-list' like in `compat-entwine', which here
is some "compat-XY.el" library when some third-party library is being
compiled and that requires that "compat-XY.el" library.

Finally, if we could not determine a version any other way, then we
know that some "compat-XY.el" library is being compiled.  In that
case, and only then, we can get the version string from
`byte-compile-current-file'.