Eagerly expand macros in compat-entwine
In order for `compat--entwine-version' to be passed on to the macro
generator functions, we cannot just bind the variable within the macro
and forget about it.  Generating a let binding instead of a
`progn'-block is likewise insufficient, as this would only affect the
evaluation, not the macro-expansion.  It is therefore necessary to
expand all macros (or at least all macros using
`compat--generate-function', which for our intent is almost every
case) while binding `compat--entwine-version', so that the binding is
preserved.

If this is not done, the version cannot be determined when loading
compat.el (the main file).  To my understanding this is not fatal,
because all this affects is that compat.elc cannot be generated, so
the slower compat.el has to be loaded.