Do not modify load-suffixes during loading
It appears that even this is too risky, and can cause issues recursive
loading (from what I understand loading the non-byte compiled
jka-compr.el causes this issue).  Instead we will check if
`compat-testing' is bound and true, in which case we explicitly load
the .el file, and otherwise keep the file without a specific suffix.