Always load prefixed definitions
We will move away from prefixed definitions later on and introduce
`compat-function' and `compat-funcall' macros. These macros make it possible to
call compatibility functions with different calling conventions. The prefixed
definitions will be deprecated. The problem with the prefixed definitions is two-fold:

1. They pollute the namespace and the API surface of Compat will become larger
and larger with every version.

2. Since Compat is not part of Emacs itself, the prefixed definitions cannot be
used by :core packages. By adding the `compat-funcall' and `compat-function'
macro to the core itself, which is a minimal addition, using compatibility
function with modified calling convention becomes possible.