compat-29: Drop broken string-pixel-width and window-pixel-width (Fix #8)
These functions seem to hangup from time to time on 28.2
(seagle0128/doom-modeline#601). By dropping the functions the hangup in
doom-modeline is resolved, since doom-modeline contains a runtime check for
string-pixel-width.

The compatibility function implementations do not contain any loops, this means
there is an underlying bug in `window-text-pixel-size' in 28.2 exposed by the
compatibility function. Either the compatibility function must be written in a
different form or we cannot provide them at all.