Revert "Add json-parse-string, json-parse-buffer, json-serialize and json-insert"
This reverts commit 7e678b3fa102a86553921d6c24056bbbe5365c7e.

I am reconsidering if we can or should provide compatibility functions for
features which may or may not be available at runtime, depending on compile
flags or the availability of a shared object and which are not only tied to the
Emacs version. The problem is that if such features are used by core packages,
e.g., python.el or jsonrpc.el, then the compatibility code for this feature
would have to be provided by Emacs itself, such that in case the feature is
missing, the compatibility code can be loaded. There is no problem for external
packages which use Compat, but it may not be a good idea to provide features as
part of Compat which are only usable by external packages. Ideally the
compatibility code from this reverted commit could be added to Emacs itself,
such that the json APIs are always available independent of libjansson.