Drop text-property-search functions
These functions are not used downstream as of now by packages which
use Compat, such that no breakage is expected. Private usage is
possible, but private users can copy text-property-search.el from the
Emacs source directly to a directory on their load-path.

text-property-search.el was added in 27.1 as a new library. The
library had bugs which were fixed in Emacs 29. We must find a coherent
solution to handle the addition of new libraries in Compat, which will
work in the long term. Every Emacs version comes with new libraries.
For example Emacs 29 alone added textsec, range, string-edit, vtable,
pixel-fill and more. It is not feasible to replicate all these
libraries in Compat.

Larger libraries are ideally distributed separately via GNU ELPA as
:core package or as a separate replica. For example seq.el is
distributed in a form, which differs from upstream seq.el. Such
compatibility packages could depend on Compat. Small libraries may be
provided as part of Compat, but we should then provide the respective
library file, such that requiring the feature will not lead to an
error. In this case we could provide text-property-search.el. The
difficulty will then be to ensure that we don't accidentally override
a newer version of the library.