Remove unnecessary decoration from NEWS.org export
For the exported text to focus on the necessary information, we strip
away everything that is not useful.  This includes the table of
contents (since we are only interested in a single heading), numbering
of heading (same reason) and the authorship.  We also remove the
"#+title" to avoid rendering a double-title at no advantage.