Revert "COMMENT out outdated sections in NEWS.org"
This reverts commit 0ec09b94d1aab6491c6e2637c6273b9a08a12615.

ELPA release mails should ideally only contain the relevant section of
the NEWS file. Unfortunately the ELPA build process does not yet give
us such a fine grained level of control. Until we implement a better
solution, we have to live with the redundant information in the mails.
Commenting out older sections of the news file has severe downsides:
Most importantly, commenting breaks rendering on GH and on the ELPA
package page. Furthermore the comment tags must be added manually and
pollute the plain text view of the file.

The plan is to mark the headline corresponding to the current release
with an :elpa-current: tag. The elpa-admin.el package build script
should then filter out the section corresponding to the marked
headline, such that only this section is sent as part of the release
announcement mail.