compat-tests: Make test for bignums more robust
As on older versions of Emacs 'bignump' will be declared by compat, the
function slot /is/ bound and we cannot rely on it being unbound to check
if the Emacs being tests supports bignums or not.  Yet as we know that
it is aliased to 'ignore', we use this fact to circumvent the issue.